to-disk: use ssh instead of exec
The exec flow is buffered and also doesn't handle things like
tty widths etc. We could replicate all of that, but it's
just way easier to fork ssh.

This feels conceptually less clean in that my preference
is actually for systems to be more autonomous, but this
way right now is the only way we could sanely get a progress
bar for example.

Signed-off-by: Colin Walters <walters@verbum.org>