libvirt: Add support for `--label`
This parallels what we do with `ephemeral`, where it's super
natural because containers have labels.

With libvirt we already have custom stuff in the XML, so
add support for reading/writing labels there, and filtering
by them in `bcvk libvirt list --label=foo` etc.

This makes it easier to use bcvk in a per-project setting.

Signed-off-by: Colin Walters <walters@verbum.org>