ephemeral: Fix `run-ssh` error handling
If we fail to launch, because we used `--rm` the container
would exit and we couldn't get its logs.

Change to removing the container image via a drop handler,
so if the system fails to launch (e.g. it's missing a kernel or bwrap)
we can get the logs from that.

Signed-off-by: Colin Walters <walters@verbum.org>